Revert "Move everything back to one file?"
This reverts commit f44415e2bde69663625558952d12c75ad6bd245e.
