+++ CMakeLists.txt
@@ -10,13 +10,13 @@ add_executable(luau-lsp
     src/JsonRpc.hpp
     src/Protocol.hpp
     src/Uri.hpp
-    # src/WorkspaceFileResolver.hpp
+    src/WorkspaceFileResolver.hpp
     src/Workspace.hpp
     src/Sourcemap.hpp
     src/TextDocument.hpp
     src/Client.hpp
     src/DocumentationParser.hpp
-    # src/LuauExt.hpp
+    src/LuauExt.hpp
     src/Utils.hpp)
 
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
