+++ src/Workspace.hpp
@@ -23,707 +23,6 @@ static std::optional<Luau::AutocompleteEntryMap> nullCallback(std::string tag, s
     return std::nullopt;
 }
 
-// Get the corresponding Luau module name for a file
-Luau::ModuleName getModuleName(const std::string& name)
-{
-    return name;
-}
-Luau::ModuleName getModuleName(const std::filesystem::path& name)
-{
-    return name.generic_string();
-}
-Luau::ModuleName getModuleName(const Uri& name)
-{
-    return name.fsPath().generic_string();
-}
-
-struct WorkspaceFileResolver
-    : Luau::FileResolver
-    , Luau::ConfigResolver
-{
-    Luau::Config defaultConfig;
-
-    // The root source node from a parsed Rojo source map
-    lsp::DocumentUri rootUri;
-    SourceNodePtr rootSourceNode;
-    mutable std::unordered_map<std::string, SourceNodePtr> realPathsToSourceNodes;
-    mutable std::unordered_map<Luau::ModuleName, SourceNodePtr> virtualPathsToSourceNodes;
-
-    // Currently opened files where content is managed by client
-    mutable std::unordered_map<Luau::ModuleName, TextDocument> managedFiles;
-    mutable std::unordered_map<std::string, Luau::Config> configCache;
-    mutable std::vector<std::pair<std::filesystem::path, std::string>> configErrors;
-
-    WorkspaceFileResolver()
-    {
-        defaultConfig.mode = Luau::Mode::Nonstrict;
-    }
-
-    /// The file is managed by the client, so FS will be out of date
-    bool isManagedFile(const Luau::ModuleName& name) const
-    {
-        return managedFiles.find(name) != managedFiles.end();
-    }
-
-    /// The name points to a virtual path (i.e., game/ or ProjectRoot/)
-    bool isVirtualPath(const Luau::ModuleName& name) const
-    {
-        return name == "game" || name == "ProjectRoot" || Luau::startsWith(name, "game/") || Luau::startsWith(name, "ProjectRoot");
-    }
-
-    std::optional<SourceNodePtr> getSourceNodeFromVirtualPath(const Luau::ModuleName& name) const
-    {
-        if (virtualPathsToSourceNodes.find(name) == virtualPathsToSourceNodes.end())
-            return std::nullopt;
-        return virtualPathsToSourceNodes.at(name);
-    }
-
-    std::optional<SourceNodePtr> getSourceNodeFromRealPath(const std::string& name) const
-    {
-        std::error_code ec;
-        auto canonicalName = std::filesystem::canonical(name, ec);
-        if (ec.value() != 0)
-            canonicalName = name;
-        auto strName = canonicalName.generic_string();
-        if (realPathsToSourceNodes.find(strName) == realPathsToSourceNodes.end())
-            return std::nullopt;
-        return realPathsToSourceNodes.at(strName);
-    }
-
-    Luau::ModuleName getVirtualPathFromSourceNode(const SourceNodePtr& sourceNode) const
-    {
-        return sourceNode->virtualPath;
-    }
-
-    std::optional<Luau::ModuleName> resolveToVirtualPath(const std::string& name) const
-    {
-        if (isVirtualPath(name))
-        {
-            return name;
-        }
-        else
-        {
-            auto sourceNode = getSourceNodeFromRealPath(name);
-            if (!sourceNode)
-                return std::nullopt;
-            return getVirtualPathFromSourceNode(sourceNode.value());
-        }
-    }
-
-    std::optional<std::filesystem::path> resolveVirtualPathToRealPath(const Luau::ModuleName& name) const
-    {
-        if (auto sourceNode = getSourceNodeFromVirtualPath(name))
-        {
-            // NOTE: this filepath is generated by the sourcemap, which is relative to the cwd where the sourcemap
-            // command was run from. Hence, we concatenate it to the end of the workspace path
-            // TODO: make sure this is correct once we make sourcemap.json generic
-            auto filePath = sourceNode.value()->getScriptFilePath();
-            if (filePath)
-                return rootUri.fsPath() / *filePath;
-            return std::nullopt;
-        }
-        return std::nullopt;
-    }
-
-    std::optional<Luau::SourceCode> readSource(const Luau::ModuleName& name) override
-    {
-        Luau::SourceCode::Type sourceType = Luau::SourceCode::Module;
-        std::optional<std::string> source;
-
-        std::filesystem::path realFileName = name;
-        if (isVirtualPath(name))
-        {
-            auto sourceNode = getSourceNodeFromVirtualPath(name);
-            if (!sourceNode)
-                return std::nullopt;
-            auto filePath = sourceNode.value()->getScriptFilePath();
-            if (!filePath)
-                return std::nullopt;
-            realFileName = filePath.value();
-            sourceType = sourceNode.value()->sourceCodeType();
-        }
-        else
-        {
-            sourceType = sourceCodeTypeFromPath(realFileName);
-        }
-
-        if (isManagedFile(name))
-        {
-            source = managedFiles.at(name).getText();
-        }
-        else
-        {
-            source = readFile(realFileName);
-            if (source && realFileName.extension() == ".json")
-            {
-                try
-                {
-                    source = "--!strict\nreturn " + jsonValueToLuau(json::parse(*source));
-                }
-                catch (const std::exception& e)
-                {
-                    // TODO: display diagnostic?
-                    std::cerr << "Failed to load JSON module: " << realFileName.generic_string() << " - " << e.what() << std::endl;
-                    return std::nullopt;
-                }
-            }
-        }
-
-        if (!source)
-            return std::nullopt;
-
-        return Luau::SourceCode{*source, sourceType};
-    }
-
-    std::optional<Luau::ModuleInfo> resolveModule(const Luau::ModuleInfo* context, Luau::AstExpr* node) override
-    {
-        // Handle require("path") for compatibility
-        if (Luau::AstExprConstantString* expr = node->as<Luau::AstExprConstantString>())
-        {
-            Luau::ModuleName name = std::string(expr->value.data, expr->value.size) + ".luau";
-            if (!readFile(name))
-            {
-                // fall back to .lua if a module with .luau doesn't exist
-                name = std::string(expr->value.data, expr->value.size) + ".lua";
-            }
-
-            return {{name}};
-        }
-        else if (Luau::AstExprGlobal* g = node->as<Luau::AstExprGlobal>())
-        {
-            if (g->name == "game")
-                return Luau::ModuleInfo{"game"};
-
-            if (g->name == "script")
-            {
-                if (auto virtualPath = resolveToVirtualPath(context->name))
-                {
-                    return Luau::ModuleInfo{virtualPath.value()};
-                }
-            }
-        }
-        else if (Luau::AstExprIndexName* i = node->as<Luau::AstExprIndexName>())
-        {
-            if (context)
-            {
-                if (strcmp(i->index.value, "Parent") == 0)
-                {
-                    // Pop the name instead
-                    auto parentPath = getParentPath(context->name);
-                    if (parentPath.has_value())
-                        return Luau::ModuleInfo{parentPath.value(), context->optional};
-                }
-
-                return Luau::ModuleInfo{context->name + '/' + i->index.value, context->optional};
-            }
-        }
-        else if (Luau::AstExprIndexExpr* i_expr = node->as<Luau::AstExprIndexExpr>())
-        {
-            if (Luau::AstExprConstantString* index = i_expr->index->as<Luau::AstExprConstantString>())
-            {
-                if (context)
-                    return Luau::ModuleInfo{context->name + '/' + std::string(index->value.data, index->value.size), context->optional};
-            }
-        }
-        else if (Luau::AstExprCall* call = node->as<Luau::AstExprCall>(); call && call->self && call->args.size >= 1 && context)
-        {
-            if (Luau::AstExprConstantString* index = call->args.data[0]->as<Luau::AstExprConstantString>())
-            {
-                Luau::AstName func = call->func->as<Luau::AstExprIndexName>()->index;
-
-                if (func == "GetService" && context->name == "game")
-                    return Luau::ModuleInfo{"game/" + std::string(index->value.data, index->value.size)};
-                if (func == "WaitForChild" || func == "FindFirstChild")
-                    if (context)
-                        return Luau::ModuleInfo{context->name + '/' + std::string(index->value.data, index->value.size), context->optional};
-            }
-        }
-
-        return std::nullopt;
-    }
-
-    std::string getHumanReadableModuleName(const Luau::ModuleName& name) const override
-    {
-        if (isVirtualPath(name))
-        {
-            if (auto realPath = resolveVirtualPathToRealPath(name))
-            {
-                return realPath->relative_path().generic_string() + " [" + name + "]";
-            }
-            else
-            {
-                return name;
-            }
-        }
-        else
-        {
-            return name;
-        }
-    }
-
-    const Luau::Config& getConfig(const Luau::ModuleName& name) const override
-    {
-        std::optional<std::filesystem::path> realPath = name;
-        if (isVirtualPath(name))
-        {
-            realPath = resolveVirtualPathToRealPath(name);
-        }
-
-        if (!realPath || !realPath->has_relative_path() || !realPath->has_parent_path())
-            return defaultConfig;
-
-        return readConfigRec(realPath->parent_path());
-    }
-
-    const Luau::Config& readConfigRec(const std::filesystem::path& path) const
-    {
-        auto it = configCache.find(path.generic_string());
-        if (it != configCache.end())
-            return it->second;
-
-        Luau::Config result = (path.has_relative_path() && path.has_parent_path()) ? readConfigRec(path.parent_path()) : defaultConfig;
-        auto configPath = path / Luau::kConfigName;
-
-        if (std::optional<std::string> contents = readFile(configPath))
-        {
-            std::optional<std::string> error = Luau::parseConfig(*contents, result);
-            if (error)
-                configErrors.push_back({configPath, *error});
-        }
-
-        return configCache[path.generic_string()] = result;
-    }
-
-    void writePathsToMap(const SourceNodePtr& node, const std::string& base)
-    {
-        node->virtualPath = base;
-        virtualPathsToSourceNodes[base] = node;
-
-        if (auto realPath = node->getScriptFilePath())
-        {
-            std::error_code ec;
-            auto canonicalName = std::filesystem::canonical(*realPath, ec);
-            if (ec.value() != 0)
-                canonicalName = *realPath;
-            realPathsToSourceNodes[canonicalName.generic_string()] = node;
-        }
-
-        for (auto& child : node->children)
-        {
-            writePathsToMap(child, base + "/" + child->name);
-        }
-    }
-
-    void updateSourceMap(const std::string& sourceMapContents)
-    {
-        realPathsToSourceNodes.clear();
-        virtualPathsToSourceNodes.clear();
-
-        try
-        {
-            auto j = json::parse(sourceMapContents);
-            rootSourceNode = std::make_shared<SourceNode>(j.get<SourceNode>());
-
-            // Write paths
-            std::string base = rootSourceNode->className == "DataModel" ? "game" : "ProjectRoot";
-            writePathsToMap(rootSourceNode, base);
-        }
-        catch (const std::exception& e)
-        {
-            // TODO: log message?
-            std::cerr << e.what() << std::endl;
-        }
-    }
-};
-
-
-namespace types
-{
-std::optional<Luau::TypeId> getTypeIdForClass(const Luau::ScopePtr& globalScope, std::optional<std::string> className)
-{
-    std::optional<Luau::TypeFun> baseType;
-    if (className.has_value())
-    {
-        baseType = globalScope->lookupType(className.value());
-    }
-    if (!baseType.has_value())
-    {
-        baseType = globalScope->lookupType("Instance");
-    }
-
-    if (baseType.has_value())
-    {
-        return baseType->type;
-    }
-    else
-    {
-        // If we reach this stage, we couldn't find the class name nor the "Instance" type
-        // This most likely means a valid definitions file was not provided
-        return std::nullopt;
-    }
-}
-
-Luau::TypeId makeLazyInstanceType(Luau::TypeArena& arena, const Luau::ScopePtr& globalScope, const SourceNodePtr& node,
-    std::optional<Luau::TypeId> parent, const WorkspaceFileResolver& fileResolver)
-{
-    Luau::LazyTypeVar ltv;
-    ltv.thunk = [&arena, globalScope, node, parent, fileResolver]()
-    {
-        // TODO: we should cache created instance types and reuse them where possible
-
-        // Look up the base class instance
-        auto baseTypeId = getTypeIdForClass(globalScope, node->className);
-        if (!baseTypeId)
-        {
-            return Luau::getSingletonTypes().anyType;
-        }
-
-        // Create the ClassTypeVar representing the instance
-        Luau::ClassTypeVar ctv{node->name, {}, baseTypeId, std::nullopt, {}, {}, "@roblox"};
-        auto typeId = arena.addType(std::move(ctv));
-
-        // Attach Parent and Children info
-        // Get the mutable version of the type var
-        if (Luau::ClassTypeVar* ctv = Luau::getMutable<Luau::ClassTypeVar>(typeId))
-        {
-
-            // Add the parent
-            if (parent.has_value())
-            {
-                ctv->props["Parent"] = Luau::makeProperty(parent.value());
-            }
-            else
-            {
-                // Search for the parent type
-                if (auto parentPath = getParentPath(node->virtualPath))
-                {
-                    if (auto parentNode = fileResolver.getSourceNodeFromVirtualPath(parentPath.value()))
-                    {
-                        ctv->props["Parent"] =
-                            Luau::makeProperty(makeLazyInstanceType(arena, globalScope, parentNode.value(), std::nullopt, fileResolver));
-                    }
-                }
-            }
-
-            // Add the children
-            for (const auto& child : node->children)
-            {
-                ctv->props[child->name] = Luau::makeProperty(makeLazyInstanceType(arena, globalScope, child, typeId, fileResolver));
-            }
-        }
-        return typeId;
-    };
-
-    return arena.addType(std::move(ltv));
-}
-
-// Magic function for `Instance:IsA("ClassName")` predicate
-std::optional<Luau::ExprResult<Luau::TypePackId>> magicFunctionInstanceIsA(
-    Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr, Luau::ExprResult<Luau::TypePackId> exprResult)
-{
-    if (expr.args.size != 1)
-        return std::nullopt;
-
-    auto index = expr.func->as<Luau::AstExprIndexName>();
-    auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-    if (!index || !str)
-        return std::nullopt;
-
-    std::optional<Luau::LValue> lvalue = tryGetLValue(*index->expr);
-    std::optional<Luau::TypeFun> tfun = scope->lookupType(std::string(str->value.data, str->value.size));
-    if (!lvalue || !tfun)
-        return std::nullopt;
-
-    Luau::TypePackId booleanPack = typeChecker.globalTypes.addTypePack({typeChecker.booleanType});
-    return Luau::ExprResult<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, tfun->type}}};
-}
-
-// Magic function for `instance:Clone()`, so that we return the exact subclass that `instance` is, rather than just a generic Instance
-static std::optional<Luau::ExprResult<Luau::TypePackId>> magicFunctionInstanceClone(
-    Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr, Luau::ExprResult<Luau::TypePackId> exprResult)
-{
-    auto index = expr.func->as<Luau::AstExprIndexName>();
-    if (!index)
-        return std::nullopt;
-
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
-    Luau::TypeId instanceType = typeChecker.checkLValueBinding(scope, *index->expr);
-    return Luau::ExprResult<Luau::TypePackId>{arena.addTypePack({instanceType})};
-}
-
-// Magic function for `Instance:FindFirstChildWhichIsA("ClassName")` and friends
-std::optional<Luau::ExprResult<Luau::TypePackId>> magicFunctionFindFirstXWhichIsA(
-    Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr, Luau::ExprResult<Luau::TypePackId> exprResult)
-{
-    if (expr.args.size < 1)
-        return std::nullopt;
-
-    auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-    if (!str)
-        return std::nullopt;
-
-    std::optional<Luau::TypeFun> tfun = scope->lookupType(std::string(str->value.data, str->value.size));
-    if (!tfun)
-        return std::nullopt;
-
-    Luau::TypeId nillableClass = Luau::makeOption(typeChecker, typeChecker.globalTypes, tfun->type);
-    return Luau::ExprResult<Luau::TypePackId>{typeChecker.globalTypes.addTypePack({nillableClass})};
-}
-
-using NameOrExpr = std::variant<std::string, Luau::AstExpr*>;
-
-// Converts a FTV and function call to a nice string
-// In the format "function NAME(args): ret"
-std::string toStringNamedFunction(Luau::ModulePtr module, const Luau::FunctionTypeVar* ftv, const NameOrExpr nameOrFuncExpr)
-{
-    Luau::ToStringOptions opts;
-    opts.functionTypeArguments = true;
-    opts.hideNamedFunctionTypeParameters = false;
-    auto functionString = Luau::toStringNamedFunction("", *ftv, opts);
-
-    // HACK: remove all instances of "_: " from the function string
-    // They don't look great, maybe we should upstream this as an option?
-    size_t index;
-    while ((index = functionString.find("_: ")) != std::string::npos)
-    {
-        functionString.replace(index, 3, "");
-    }
-
-    // If a name has already been provided, just use that
-    if (auto name = std::get_if<std::string>(&nameOrFuncExpr))
-    {
-        return "function " + *name + functionString;
-    }
-    auto funcExprPtr = std::get_if<Luau::AstExpr*>(&nameOrFuncExpr);
-
-    // TODO: error here?
-    if (!funcExprPtr)
-        return "function" + functionString;
-    auto funcExpr = *funcExprPtr;
-
-    // See if its just in the form `func(args)`
-    if (auto local = funcExpr->as<Luau::AstExprLocal>())
-    {
-        return "function " + std::string(local->local->name.value) + functionString;
-    }
-    else if (auto global = funcExpr->as<Luau::AstExprGlobal>())
-    {
-        return "function " + std::string(global->name.value) + functionString;
-    }
-    else if (funcExpr->as<Luau::AstExprGroup>() || funcExpr->as<Luau::AstExprFunction>())
-    {
-        // In the form (expr)(args), which implies thats its probably a IIFE
-        return "function" + functionString;
-    }
-
-    // See if the name belongs to a ClassTypeVar
-    Luau::TypeId* parentIt = nullptr;
-    std::string methodName;
-    std::string baseName;
-
-    if (auto indexName = funcExpr->as<Luau::AstExprIndexName>())
-    {
-        parentIt = module->astTypes.find(indexName->expr);
-        methodName = std::string(1, indexName->op) + indexName->index.value;
-        // If we are calling this as a method ':', we should implicitly hide self
-        opts.hideFunctionSelfArgument = indexName->op == ':';
-        // We can try and give a temporary base name from what we can infer by the index, and then attempt to improve it with proper information
-        baseName = Luau::toString(indexName->expr);
-        trim(baseName); // Trim it, because toString is probably not meant to be used in this context (it has whitespace)
-    }
-    else if (auto indexExpr = funcExpr->as<Luau::AstExprIndexExpr>())
-    {
-        parentIt = module->astTypes.find(indexExpr->expr);
-        methodName = Luau::toString(indexExpr->index);
-        // We can try and give a temporary base name from what we can infer by the index, and then attempt to improve it with proper information
-        baseName = Luau::toString(indexName->expr);
-        trim(baseName); // Trim it, because toString is probably not meant to be used in this context (it has whitespace)
-    }
-
-    if (parentIt)
-    {
-        Luau::TypeId parentType = Luau::follow(*parentIt);
-        if (auto typeName = Luau::getName(parentType))
-        {
-            baseName = *typeName;
-        }
-        else if (auto parentClass = Luau::get<Luau::ClassTypeVar>(parentType))
-        {
-            baseName = parentClass->name;
-        }
-        // if (auto parentUnion = Luau::get<UnionTypeVar>(parentType))
-        // {
-        //     return returnFirstNonnullOptionOfType<ClassTypeVar>(parentUnion);
-        // }
-    }
-    else
-    {
-        // TODO: anymore we can do?
-        baseName = "_";
-    }
-
-    return "function " + baseName + methodName + functionString;
-}
-
-// Duplicated from Luau/TypeInfer.h, since its static
-std::optional<Luau::AstExpr*> matchRequire(const Luau::AstExprCall& call)
-{
-    const char* require = "require";
-
-    if (call.args.size != 1)
-        return std::nullopt;
-
-    const Luau::AstExprGlobal* funcAsGlobal = call.func->as<Luau::AstExprGlobal>();
-    if (!funcAsGlobal || funcAsGlobal->name != require)
-        return std::nullopt;
-
-    if (call.args.size != 1)
-        return std::nullopt;
-
-    return call.args.data[0];
-}
-
-
-} // namespace types
-
-// TODO: should upstream this
-struct FindNodeType : public Luau::AstVisitor
-{
-    const Luau::Position pos;
-    const Luau::Position documentEnd;
-    Luau::AstNode* best = nullptr;
-
-    explicit FindNodeType(Luau::Position pos, Luau::Position documentEnd)
-        : pos(pos)
-        , documentEnd(documentEnd)
-    {
-    }
-
-    bool visit(Luau::AstNode* node) override
-    {
-        if (node->location.contains(pos))
-        {
-            best = node;
-            return true;
-        }
-
-        // Edge case: If we ask for the node at the position that is the very end of the document
-        // return the innermost AST element that ends at that position.
-
-        if (node->location.end == documentEnd && pos >= documentEnd)
-        {
-            best = node;
-            return true;
-        }
-
-        return false;
-    }
-
-    virtual bool visit(class Luau::AstType* node) override
-    {
-        return visit(static_cast<Luau::AstNode*>(node));
-    }
-
-    bool visit(Luau::AstStatBlock* block) override
-    {
-        visit(static_cast<Luau::AstNode*>(block));
-
-        for (Luau::AstStat* stat : block->body)
-        {
-            if (stat->location.end < pos)
-                continue;
-            if (stat->location.begin > pos)
-                break;
-
-            stat->visit(this);
-        }
-
-        return false;
-    }
-};
-
-Luau::AstNode* findNodeOrTypeAtPosition(const Luau::SourceModule& source, Luau::Position pos)
-{
-    const Luau::Position end = source.root->location.end;
-    if (pos < source.root->location.begin)
-        return source.root;
-
-    if (pos > end)
-        pos = end;
-
-    FindNodeType findNode{pos, end};
-    findNode.visit(source.root);
-    return findNode.best;
-}
-
-std::optional<Luau::Location> lookupTypeLocation(const Luau::Scope& deepScope, const Luau::Name& name)
-{
-    const Luau::Scope* scope = &deepScope;
-    while (true)
-    {
-        auto it = scope->typeAliasLocations.find(name);
-        if (it != scope->typeAliasLocations.end())
-            return it->second;
-
-        if (scope->parent)
-            scope = scope->parent.get();
-        else
-            return std::nullopt;
-    }
-}
-
-std::optional<Luau::Property> lookupProp(const Luau::TypeId& parentType, const Luau::Name& name)
-{
-    if (auto ctv = Luau::get<Luau::ClassTypeVar>(parentType))
-    {
-        if (auto prop = Luau::lookupClassProp(ctv, name))
-            return *prop;
-    }
-    else if (auto tbl = Luau::get<Luau::TableTypeVar>(parentType))
-    {
-        if (tbl->props.find(name) != tbl->props.end())
-        {
-            return tbl->props.at(name);
-        }
-    }
-    else if (auto mt = Luau::get<Luau::MetatableTypeVar>(parentType))
-    {
-        if (auto tbl = Luau::get<Luau::TableTypeVar>(mt->table))
-        {
-            if (tbl->props.find(name) != tbl->props.end())
-            {
-                return tbl->props.at(name);
-            }
-        }
-
-        // TODO: we should respect metatable __index
-    }
-    // else if (auto i = get<Luau::IntersectionTypeVar>(parentType))
-    // {
-    //     for (Luau::TypeId ty : i->parts)
-    //     {
-    //         // TODO: find the corresponding ty
-    //     }
-    // }
-    // else if (auto u = get<Luau::UnionTypeVar>(parentType))
-    // {
-    //     // Find the corresponding ty
-    // }
-    return std::nullopt;
-}
-
-Luau::Position convertPosition(const lsp::Position& position)
-{
-    LUAU_ASSERT(position.line <= UINT_MAX);
-    LUAU_ASSERT(position.character <= UINT_MAX);
-    return Luau::Position{static_cast<unsigned int>(position.line), static_cast<unsigned int>(position.character)};
-}
-
-lsp::Position convertPosition(const Luau::Position& position)
-{
-    return lsp::Position{static_cast<size_t>(position.line), static_cast<size_t>(position.column)};
-}
-
 class WorkspaceFolder
 {
 public:
